refactor: remove unnecessary comments and format some stuff
