--- src/main/java/se/datasektionen/calypso/auth/entities/permissions/Permission.java
@@ -1,12 +1,0 @@
-package se.datasektionen.calypso.auth.entities.permissions;
-
-import lombok.Data;
-
-@Data
-public class Permission {
-
-	private String id;
-
-	private String scope;
-}
-//TODO might remove
